ref: Only LinkAction is a scriptable object
