--- unity/CAVE/Assets/Resources/Scripts/Actions/LinkAction.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b24652ad07205a6469e8672142bc14c3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
