--- unity/CAVE/Assets/Resources/Scripts/Actions/ObjectAction.cs
@@ -1,13 +1,0 @@
-using System;
-
-using UnityEngine;
-
-namespace Writing3D
-{
-    [CreateAssetMenu(fileName = "ObjectAction", menuName = "W3D/Action/ObjectAction", order = 1)]
-    [Serializable]
-    public class ObjectAction : Action
-    {
-        [SerializeField] public Transition Transition;
-    }
-}
