--- unity/CAVE/Assets/Resources/Scripts/Transitions/VisibleTransition.cs
@@ -1,13 +1,0 @@
-using System;
-
-using UnityEngine;
-
-namespace Writing3D
-{
-    [CreateAssetMenu(fileName = "VisibleTransition", menuName = "W3D/Transition/VisibleTransition", order = 1)]
-    [Serializable]
-    public class VisibleTransition : Transition
-    {
-        [SerializeField] public bool Visible;
-    }
-}
