--- unity/CAVE/Assets/Resources/Scripts/Transitions/VisibleTransition.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: abdece1e303680647a65d4027c1519b3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
